requirements.txt
# deepspeed==0.18.4
toml
transformers
diffusers>=0.35.1
datasets
pillow
sentencepiece
protobuf
peft
torch-optimi
tensorboard
tqdm
safetensors
# bitsandbytes
imageio[ffmpeg]
av
einops
accelerate
loguru
# flash-attn==2.8.1
omegaconf
iopath
termcolor
hydra-core
easydict
ftfy
pytorch-optimizer
wandb
optimum-quanto
psutil
mpi4py
wheel
ninja
pynvml
sentry_sdk
# Tools Dependencies
opencv-python
imagehash
google-generativeai
python-dotenv
qwen-vl-utils